services/backend/pyproject.toml
[project]
name = "projectsyte"
version = "0.1.0"
description = ""
authors = [
    {name = "Maxutka73",email = "gouynoran@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14,<4.0"
dependencies = [
    "annotated-doc (==0.0.3)",
    "annotated-types (==0.7.0)",
    "anyio (==4.11.0)",
    "build (==1.3.0)",
    "cachecontrol (==0.14.4)",
    "certifi (==2025.11.12)",
    "charset-normalizer (==3.4.4)",
    "cleo (==2.1.0)",
    "click (==8.3.0)",
    "colorama (==0.4.6)",
    "crashtest (==0.4.1)",
    "distlib (==0.4.0)",
    "dulwich (==0.24.10)",
    "exceptiongroup (==1.3.0)",
    "fastapi (==0.121.0)",
    "fastjsonschema (==2.21.2)",
    "filelock (==3.20.0)",
    "findpython (==0.7.1)",
    "h11 (==0.16.0)",
    "httpcore (==1.0.9)",
    "httpx (==0.28.1)",
    "idna (==3.11)",
    "installer (==0.7.0)",
    "jaraco-classes (==3.4.0)",
    "jaraco-context (==6.0.1)",
    "jaraco-functools (==4.3.0)",
    "keyring (==25.7.0)",
    "more-itertools (==10.8.0)",
    "msgpack (==1.1.2)",
    "packaging (==25.0)",
    "pbs-installer (==2025.12.5)",
    "pillow (==12.0.0)",
    "pkginfo (==1.12.1.2)",
    "platformdirs (==4.5.1)",
    "poetry (==2.2.1)",
    "poetry-core (==2.2.1)",
    "psutil (==7.2.0)",
    "psycopg (==3.3.2)",
    "psycopg-pool (==3.3.0)",
    "pydantic (==2.12.4)",
    "pydantic-settings (==2.11.0)",
    "pydantic-core (==2.41.5)",
    "pyproject-hooks (==1.2.0)",
    "python-dotenv (==1.2.1)",
    "python-multipart (==0.0.20)",
    "pywin32-ctypes (==0.2.3)",
    "rapidfuzz (==3.14.3)",
    "requests (==2.32.5)",
    "requests-toolbelt (==1.0.0)",
    "shellingham (==1.5.4)",
    "sniffio (==1.3.1)",
    "starlette (==0.49.3)",
    "tomlkit (==0.13.3)",
    "trove-classifiers (==2025.12.1.14)",
    "typing-inspection (==0.4.2)",
    "typing-extensions (==4.15.0)",
    "tzdata (==2025.3)",
    "urllib3 (==2.6.2)",
    "virtualenv (==20.35.4)",
    "watchfiles (==1.1.1)",
    "zstandard (==0.25.0)",
    "uvicorn[standard] (==0.30.6)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
